+++ test/dag_benchmarks.cc
@@ -256,34 +256,37 @@ class restartable_thread {
 	}
 };
 
-class benchmark_scheduler final : public abstract_scheduler {
+namespace celerity::detail {
+
+class test_benchmark_scheduler : public scheduler {
   public:
-	benchmark_scheduler(restartable_thread& thread, const size_t num_nodes, const node_id local_node_id, const system_info& system_info,
-	    abstract_scheduler::delegate* const delegate, command_recorder* const crec, instruction_recorder* const irec)
-	    : abstract_scheduler(num_nodes, local_node_id, system_info, delegate, crec, irec), m_thread(&thread) {
-		m_thread->start([this] { schedule(); });
+	test_benchmark_scheduler(restartable_thread& thread, const size_t num_nodes, const node_id local_node_id, const system_info& system_info,
+	    scheduler::delegate* const delegate, command_recorder* const crec, instruction_recorder* const irec)
+	    : scheduler(scheduler::test_start_idle_tag{}, num_nodes, local_node_id, system_info, delegate, crec, irec), m_thread(&thread) //
+	{
+		m_thread->start([this] { test_invoke_thread_main(); });
 	}
 
-	benchmark_scheduler(const benchmark_scheduler&) = delete;
-	benchmark_scheduler(benchmark_scheduler&&) = delete;
-	benchmark_scheduler& operator=(const benchmark_scheduler&) = delete;
-	benchmark_scheduler& operator=(benchmark_scheduler&&) = delete;
+	test_benchmark_scheduler(const test_benchmark_scheduler&) = delete;
+	test_benchmark_scheduler(test_benchmark_scheduler&&) = delete;
+	test_benchmark_scheduler& operator=(const test_benchmark_scheduler&) = delete;
+	test_benchmark_scheduler& operator=(test_benchmark_scheduler&&) = delete;
 
-	void join() {
+	~test_benchmark_scheduler() {
 		// schedule() will exit as soon as it has acknowledged the shutdown epoch
 		m_thread->join();
 	}
-
-	~benchmark_scheduler() override { join(); }
 
   private:
 	restartable_thread* m_thread;
 };
 
+} // namespace celerity::detail
+
 struct scheduler_benchmark_context : private task_manager::delegate { // NOLINT(cppcoreguidelines-virtual-class-destructor)
 	const size_t num_nodes;
-	benchmark_scheduler schdlr;
 	task_manager tm{num_nodes, nullptr, this, benchmark_task_manager_policy};
+	detail::test_benchmark_scheduler schdlr;
 	test_utils::mock_buffer_factory mbf;
 
 	explicit scheduler_benchmark_context(restartable_thread& thrd, const size_t num_nodes, const size_t num_devices_per_node)
@@ -303,7 +306,6 @@ struct scheduler_benchmark_context : private task_manager::delegate { // NOLINT(
 		const auto tid = tm.generate_epoch_task(celerity::detail::epoch_action::shutdown);
 		// There is no executor thread and notifications are processed in-order, so we can immediately notify the scheduler about shutdown-epoch completion
 		schdlr.notify_epoch_reached(tid);
-		schdlr.join(); // must join explicitly, since `tm` is destroyed before `schdlr` and `schdlr` references task ids kept alive by `tm`
 	}
 
 	template <int KernelDims, typename CGF>
@@ -345,6 +347,8 @@ template <typename BenchmarkContext>
 [[gnu::noinline]] BenchmarkContext&& generate_soup_graph(BenchmarkContext&& ctx, const size_t num_tasks) {
 	test_utils::mock_buffer<2> buf = ctx.mbf.create_buffer(range<2>{ctx.num_nodes, num_tasks}, true /* host_initialized */);
 	for(size_t t = 0; t < num_tasks; ++t) {
+		ctx.create_task(range<1>{ctx.num_nodes},
+		    [&](handler& cgh) { buf.get_access<access_mode::read_write>(cgh, [=](chunk<1> ck) { return subrange<2>{{ck.offset[0], t}, {ck.range[0], 1}}; }); });
 		ctx.create_task(range<1>{ctx.num_nodes},
 		    [&](handler& cgh) { buf.get_access<access_mode::read_write>(cgh, [=](chunk<1> ck) { return subrange<2>{{ck.offset[0], t}, {ck.range[0], 1}}; }); });
 	}